temporary commit

+++ src/main/java/com/company/generators/BinaryGenerator.java
@@ -2,5 +2,5 @@ package com.company.generators;
 
 public interface BinaryGenerator {
     boolean hasNext();
-    int[] next();
+    long[] next();
 }
