+++ src/main/java/com/company/generators/utils/Translator.java
@@ -1,9 +1,9 @@
 package com.company.generators.utils;
 
 public class Translator {
-    public static int[] translateToArrayOfBits(int number,int totalSize){
-        int[] bits = new int[totalSize];
-        int lowerBit = 1;
+    public static long[] translateToArrayOfBits(long number,int totalSize){
+        long[] bits = new long[totalSize];
+        long lowerBit = 1;
         for (int i = totalSize - 1; i >= 0; i--) {
             bits[i] = number & lowerBit;
             number = number >> 1;
